Added performance tests and Test Plans.
