+++ GeoJsonUtilsApp/GeoJsonUtilsAppTests/Performance.xctestplan
@@ -1,0 +1,28 @@
+{
+  "configurations" : [
+    {
+      "id" : "23CB7A39-FF17-42EF-AC1D-2EAFBF167A45",
+      "name" : "Configuration 1",
+      "options" : {
+
+      }
+    }
+  ],
+  "defaultOptions" : {
+
+  },
+  "testTargets" : [
+    {
+      "skippedTests" : [
+        "GeoJsonUtilsAppTests",
+        "MapKitToolsTests"
+      ],
+      "target" : {
+        "containerPath" : "container:GeoJsonUtilsApp.xcodeproj",
+        "identifier" : "803E504B21FF6C4F00699E12",
+        "name" : "GeoJsonUtilsAppTests"
+      }
+    }
+  ],
+  "version" : 1
+}
