+++ GeoJsonUtilsApp/GeoJsonUtilsAppTests/UnitTests.xctestplan
@@ -1,0 +1,32 @@
+{
+  "configurations" : [
+    {
+      "id" : "DB0C12D8-925B-4BEC-A5EE-BD1DB10E2366",
+      "name" : "Configuration 1",
+      "options" : {
+
+      }
+    }
+  ],
+  "defaultOptions" : {
+    "mainThreadCheckerEnabled" : false,
+    "targetForVariableExpansion" : {
+      "containerPath" : "container:GeoJsonUtilsApp.xcodeproj",
+      "identifier" : "8026723A21F97F3E00F79328",
+      "name" : "GeoJsonUtilsApp"
+    }
+  },
+  "testTargets" : [
+    {
+      "skippedTests" : [
+        "PerformanceTests"
+      ],
+      "target" : {
+        "containerPath" : "container:GeoJsonUtilsApp.xcodeproj",
+        "identifier" : "803E504B21FF6C4F00699E12",
+        "name" : "GeoJsonUtilsAppTests"
+      }
+    }
+  ],
+  "version" : 1
+}
